Update Grid.tsx
Added in some more color square code. When you click on a square, we check colors of squares left, right. top, down and cycle to the next color. Can make some fun patterns this way. Initializing random colors is alright, but not as fun as starting with the same color.