Added two more shapes and options area to ui
Added the square and triangle shape and included an options area to select the shape. Grid is growing, I'll want to remove the shapePropagation method to its own service. I also want to add in the randomizerStepSequence to create a service that propagates shapes based on random and patterned clicks into the grid. Would be controlled programatically.