Tracking click events to rerun sequences
We are now tracking click events so we can rerun sequences. Currently this is done when you reset the grid. I programmed in two different kind of rendering loops, and an internal switch that will rerun the sequence as the original shapes, or as the currently selected shape. It's pretty cool to run as currently selected shapes since you can see click patterns that would have been hard to do otherwise.